Require password only if locked and adding an account
